Require all images to be listed in book file
xelatex running in a Docker container does not have access to images in
the working directory that are available when the build is done locally.
So we need to copy them to tmpdir that is mounted into the container.

Unknown files are assumed to be images and are mirrored as is to the
appropriate location in the temporary directory.